refactor: rename issues -> fixes
