+++ tester/src/test/scala/fixes/Issue963.scala
@@ -1,4 +1,4 @@
-package issues
+package fixes
 
 import org.scalatest.funsuite.AnyFunSuite
